requirements.txt
# Budget Tracker Agent - Requirements
# Python 3.7+

# Google Gemini integration for natural language processing
google-generativeai>=0.3.0

# Environment variable management
python-dotenv>=1.0.0

# MongoDB for Long-Term Memory
pymongo>=4.0.0

# FastAPI for REST API backend
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.0.0

# Core dependencies (standard library - no external packages needed for core functionality)
# - json
# - os
# - statistics
# - datetime
# - abc
# - typing
# - uuid
# - argparse

